refactor: Improve code reliability (#60)

+++ edgesql_kaggle.py
@@ -1,4 +1,3 @@
-#from kaggle.api.kaggle_api_extended import KaggleApi
 import pandas as pd
 import os
 import zipfile
@@ -16,18 +15,16 @@ class EdgSQLKaggle:
     def __init__(self):
         """
         Initialize EdgSQLKaggle with Kaggle API credentials.
-
-        Args:
-            username (str): Kaggle username.
-            api_key (str): Kaggle API key.
         """
-
         self._load_credentials()
         self.api = None
         self._authenticate_kaggle()
+        self._df = None
 
     def _load_credentials(self):
-        # Check if kaggle.json file is available
+        """
+        Load Kaggle credentials from kaggle.json or environment variables.
+        """
         kaggle_json_path = os.path.expanduser('~/.kaggle/kaggle.json')
         if os.path.isfile(kaggle_json_path):
             with open(kaggle_json_path, 'r') as f:
@@ -36,24 +33,19 @@ class EdgSQLKaggle:
                 kaggle_key = kaggle_json_data.get('key')
 
                 if kaggle_username and kaggle_key:
-                    # Set the environment variables
                     os.environ['KAGGLE_USERNAME'] = kaggle_username
                     os.environ['KAGGLE_KEY'] = kaggle_key
                     self.username = kaggle_username
                     self.api_key = kaggle_key
-        elif os.environ['KAGGLE_USERNAME'] != '__NONE__' \
-            and os.environ['KAGGLE_KEY'] != '__NONE__':
-            self.username = os.environ.get('KAGGLE_USERNAME')
-            self.api_key = os.environ.get('KAGGLE_KEY')
+        elif 'KAGGLE_USERNAME' in os.environ and 'KAGGLE_KEY' in os.environ:
+            self.username = os.environ['KAGGLE_USERNAME']
+            self.api_key = os.environ['KAGGLE_KEY']
         else:
-            raise Exception("Kaggle credentials not found.")
+            raise RuntimeError("Kaggle credentials not found.")
 
     def _authenticate_kaggle(self):
         """
         Authenticate with Kaggle API using provided credentials.
-
-        Returns:
-            bool: True if authentication is successful, False otherwise.
         """
         try:
             configuration = Configuration()
@@ -62,10 +54,8 @@ class EdgSQLKaggle:
 
             self.api = kaggle.KaggleApi(kaggle.ApiClient(configuration))
             self.api.authenticate()
-
-            return True
         except ApiException as e:
-            raise Exception(f'Error authenticating with Kaggle API: {e}')
+            raise RuntimeError(f'Error authenticating with Kaggle API: {e}') from e
 
     def get_dataset(self):
         """
@@ -74,7 +64,7 @@ class EdgSQLKaggle:
         Returns:
             pd.DataFrame or None: The dataset if available, None otherwise.
         """
-        return self._df if hasattr(self, '_df') else None
+        return self._df
 
     def import_dataset(self, dataset_name, data_file):
         """
@@ -82,6 +72,7 @@ class EdgSQLKaggle:
 
         Args:
             dataset_name (str): The name of the dataset on Kaggle to be downloaded.
+            data_file (str): The name of the file within the dataset to be loaded as a DataFrame.
 
         Returns:
             bool: True if the import is successful, False otherwise.
@@ -89,37 +80,33 @@ class EdgSQLKaggle:
         Raises:
             Exception: If any error occurs during the download or import process.
         """
-
         if not isinstance(dataset_name, str):
-            raise Exception('Error: dataset_name must be strings.')
-
+            raise ValueError('Error: dataset_name must be a string.')
+        if not isinstance(data_file, str):
+            raise ValueError('Error: data_file must be a string.')
         if not self.api:
-            raise Exception('Error: Kaggle API not authenticated.')
+            raise RuntimeError('Error: Kaggle API not authenticated.')
 
         try:
             # Download the Kaggle dataset
             self.api.dataset_download_file(dataset_name, data_file)
-
-            # Check if the downloaded file is a zip file
-            zip_file = data_file+'.zip'
+            zip_file = data_file + '.zip'
             if zipfile.is_zipfile(zip_file):
                 with zipfile.ZipFile(zip_file, 'r') as zip_ref:
-                    zip_ref.extractall('.')  # Extract the contents of the zip file
-                os.remove(zip_file)  # Remove the downloaded zip file
-
+                    zip_ref.extractall('.')  
+                os.remove(zip_file)
             self._df = pd.read_csv(data_file)
 
             # Delete temporary files
             os.remove(data_file)
-
-            return True  # Indicating successful import
+            return True
         except ApiException as e:
-            raise Exception(f'Error importing Kaggle dataset "{dataset_name}": {e}')
-        except FileNotFoundError:
-            raise Exception(f'Error: Dataset "{dataset_name}" not found on Kaggle.')
-        except pd.errors.EmptyDataError:
-            raise Exception(f'Error: Dataset "{dataset_name}" is empty or contains no data.')
+            raise RuntimeError(f'Error importing Kaggle dataset "{dataset_name}": {e}') from e
+        except FileNotFoundError as e:
+            raise FileNotFoundError(f'Error: Dataset "{dataset_name}" not found on Kaggle.') from e
+        except pd.errors.EmptyDataError as e:
+            raise ValueError(f'Error: Dataset "{dataset_name}" is empty or contains no data.') from e
         except pd.errors.ParserError as e:
-            raise Exception(f'Error parsing dataset "{dataset_name}": {e}')
+            raise ValueError(f'Error parsing dataset "{dataset_name}": {e}') from e
         except Exception as e:
-            raise Exception(f'Error importing Kaggle dataset "{dataset_name}": {e}')
+            raise RuntimeError(f'Error importing Kaggle dataset "{dataset_name}": {e}') from e
